Improve code organization.
