+++ politeiawww/cmds.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2021-2022 The Decred developers
+// Copyright (c) 2022 The Decred developers
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
@@ -7,10 +7,8 @@ package main
 import (
 	"context"
 	"database/sql"
-	"encoding/json"
 	"fmt"
 	"regexp"
-	"strings"
 
 	plugin "github.com/decred/politeia/politeiawww/plugin/v1"
 	"github.com/decred/politeia/politeiawww/user"
@@ -18,15 +16,15 @@ import (
 	"github.com/pkg/errors"
 )
 
-// pluginNewUser executes a plugin command that results in the creation of a
-// new user in the user database.
+// newUserCmd executes a plugin command that results in the creation of a new
+// user in the user database.
 //
 // Any updates made to the session will be persisted by the caller.
 //
 // This function assumes the caller has verified that the plugin command is
 // for the user plugin.
-func (p *politeiawww) pluginNewUser(ctx context.Context, session *plugin.Session, cmd plugin.Cmd) (*plugin.Reply, error) {
-	log.Tracef("NewUser: %v %v %v", cmd.PluginID, cmd.Cmd, session.UserID)
+func (p *politeiawww) newUserCmd(ctx context.Context, session *plugin.Session, cmd plugin.Cmd) (*plugin.Reply, error) {
+	log.Tracef("newUserCmd: %v %v %v", cmd.PluginID, cmd.Cmd, session.UserID)
 
 	// Setup the database transaction
 	tx, cancel, err := p.beginTx()
@@ -52,12 +50,12 @@ func (p *politeiawww) pluginNewUser(ctx context.Context, session *plugin.Session
 	}
 
 	// Execute the pre plugin hooks
-	reply, err = p.pluginPreHooks(tx, usr,
+	reply, err = p.preHooks(tx, usr,
 		plugin.HookArgs{
 			Type:  plugin.HookPreNewUser,
 			Cmd:   cmd,
 			Reply: nil,
-			User:  nil, // User is set by pluginHooks
+			User:  nil, // User is set by hook()
 		})
 	if err != nil {
 		return nil, err
@@ -87,12 +85,12 @@ func (p *politeiawww) pluginNewUser(ctx context.Context, session *plugin.Session
 	updateUser(usr, pluginUser, cmd.PluginID)
 
 	// Execute the post plugin hooks
-	err = p.pluginPostHooks(tx, usr,
+	err = p.postHooks(tx, usr,
 		plugin.HookArgs{
 			Type:  plugin.HookPostNewUser,
 			Cmd:   cmd,
 			Reply: reply,
-			User:  nil, // User is set by pluginHooks
+			User:  nil, // User is set by hook()
 		})
 	if err != nil {
 		return nil, err
@@ -118,14 +116,14 @@ func (p *politeiawww) pluginNewUser(ctx context.Context, session *plugin.Session
 	return reply, nil
 }
 
-// pluginWrite executes a plugin command that writes data.
+// writeCmd executes a plugin command that writes data.
 //
 // Any updates made to the session will be persisted by the caller.
 //
 // This function assumes the caller has verified that the plugin is a
 // registered plugin.
-func (p *politeiawww) pluginWrite(ctx context.Context, session *plugin.Session, cmd plugin.Cmd) (*plugin.Reply, error) {
-	log.Tracef("pluginWrite: %v %v %v", cmd.PluginID, cmd.Cmd, session.UserID)
+func (p *politeiawww) writeCmd(ctx context.Context, session *plugin.Session, cmd plugin.Cmd) (*plugin.Reply, error) {
+	log.Tracef("writeCmd: %v %v %v", cmd.PluginID, cmd.Cmd, session.UserID)
 
 	// Setup the database transaction
 	tx, cancel, err := p.beginTx()
@@ -157,12 +155,12 @@ func (p *politeiawww) pluginWrite(ctx context.Context, session *plugin.Session, 
 	}
 
 	// Execute the pre plugin hooks
-	reply, err = p.pluginPreHooks(tx, usr,
+	reply, err = p.preHooks(tx, usr,
 		plugin.HookArgs{
 			Type:  plugin.HookPreWrite,
 			Cmd:   cmd,
 			Reply: nil,
-			User:  nil, // User is set by pluginHooks
+			User:  nil, // User is set by hook()
 		})
 	if err != nil {
 		return nil, err
@@ -197,12 +195,12 @@ func (p *politeiawww) pluginWrite(ctx context.Context, session *plugin.Session, 
 	updateUser(usr, pluginUser, cmd.PluginID)
 
 	// Execute the post plugin hooks
-	err = p.pluginPostHooks(tx, usr,
+	err = p.postHooks(tx, usr,
 		plugin.HookArgs{
 			Type:  plugin.HookPostWrite,
 			Cmd:   cmd,
 			Reply: reply,
-			User:  nil, // User is set by pluginHooks
+			User:  nil, // User is set by hook()
 		})
 	if err != nil {
 		return nil, err
@@ -231,15 +229,15 @@ func (p *politeiawww) pluginWrite(ctx context.Context, session *plugin.Session, 
 	return reply, nil
 }
 
-// pluginRead executes a read-only plugin command. The read operation is not
+// readCmd executes a read-only plugin command. The read operation is not
 // atomic.
 //
 // Any updates made to the session will be persisted by the caller.
 //
 // This function assumes the caller has verified that the plugin is a
 // registered plugin.
-func (p *politeiawww) pluginRead(ctx context.Context, session *plugin.Session, cmd plugin.Cmd) (*plugin.Reply, error) {
-	log.Tracef("pluginRead: %v %v %v", cmd.PluginID, cmd.Cmd, session.UserID)
+func (p *politeiawww) readCmd(ctx context.Context, session *plugin.Session, cmd plugin.Cmd) (*plugin.Reply, error) {
+	log.Tracef("readCmd: %v %v %v", cmd.PluginID, cmd.Cmd, session.UserID)
 
 	// Get the user. A session user may or may not exist.
 	var (
@@ -285,14 +283,14 @@ func (p *politeiawww) pluginRead(ctx context.Context, session *plugin.Session, c
 	return reply, nil
 }
 
-// pluginPreHooks executes the provided pre hook for all plugins. Pre hooks are
-// used to perform validation on the plugin command.
+// preHooks executes the provided pre hook for all plugins. Pre hooks are used
+// to perform validation on the plugin command.
 //
 // A plugin reply will be returned if one of the plugins throws a user error
 // during hook execution. The user error will be embedded in the plugin
 // reply. Unexpected errors result in a standard golang error being returned.
-func (p *politeiawww) pluginPreHooks(tx *sql.Tx, usr *user.User, h plugin.HookArgs) (*plugin.Reply, error) {
-	err := p.pluginHooks(tx, h, usr)
+func (p *politeiawww) preHooks(tx *sql.Tx, usr *user.User, h plugin.HookArgs) (*plugin.Reply, error) {
+	err := p.hook(tx, h, usr)
 	if err != nil {
 		var ue plugin.UserError
 		if errors.As(err, &ue) {
@@ -305,18 +303,18 @@ func (p *politeiawww) pluginPreHooks(tx *sql.Tx, usr *user.User, h plugin.HookAr
 	return nil, nil
 }
 
-// pluginPostHooks executes the provided post hook for all user plugins.
+// postHooks executes the provided post hook for all user plugins.
 //
 // Post hooks are not able to throw plugin errors like the pre hooks are. Any
 // error returned by a plugin from a post hook will be treated as an unexpected
 // error.
-func (p *politeiawww) pluginPostHooks(tx *sql.Tx, usr *user.User, h plugin.HookArgs) error {
-	return p.pluginHooks(tx, h, usr)
+func (p *politeiawww) postHooks(tx *sql.Tx, usr *user.User, h plugin.HookArgs) error {
+	return p.hook(tx, h, usr)
 }
 
-// pluginHooks executes a hook on on all plugins. A sql Tx may or may not exist
+// hook executes a hook on on all plugins. A sql Tx may or may not exist
 // depending on the whether the caller is executing an atomic operation.
-func (p *politeiawww) pluginHooks(tx *sql.Tx, h plugin.HookArgs, usr *user.User) error {
+func (p *politeiawww) hook(tx *sql.Tx, h plugin.HookArgs, usr *user.User) error {
 	for _, pluginID := range p.pluginIDs {
 		// Get the plugin
 		p, ok := p.plugins[pluginID]
@@ -415,81 +413,3 @@ var (
 	// pluginID,key,["value1","value2"] matches ["value1","value2"]
 	regexpPluginSettingMulti = regexp.MustCompile(`(\[.*\]$)`)
 )
-
-// parsePluginSetting parses a plugin setting. Plugin settings will be in
-// following format. The value may be a single value or an array of values.
-//
-// pluginID,key,value
-// pluginID,key,["value1","value2","value3"...]
-//
-// When multiple values are provided, the values must be formatted as a JSON
-// encoded []string. Both of the following JSON formats are acceptable.
-//
-// pluginID,key,["value1","value2","value3"]
-// pluginsetting="pluginID,key,[\"value1\",\"value2\",\"value3\"]"
-func parsePluginSetting(setting string) (string, *plugin.Setting, error) {
-	formatMsg := `expected plugin setting format is ` +
-		`pluginID,key,value OR pluginID,key,["value1","value2","value3"]`
-
-	// Parse the plugin setting
-	var (
-		parsed = strings.Split(setting, ",")
-
-		// isMulti indicates whether the plugin setting contains
-		// multiple values. If the setting only contains a single
-		// value then isMulti will be false.
-		isMulti = regexpPluginSettingMulti.MatchString(setting)
-	)
-	switch {
-	case len(parsed) < 3:
-		return "", nil, errors.Errorf("missing csv entry '%v'; %v",
-			setting, formatMsg)
-	case len(parsed) == 3:
-		// This is expected; continue
-	case len(parsed) > 3 && isMulti:
-		// This is expected; continue
-	default:
-		return "", nil, errors.Errorf("invalid format '%v'; %v",
-			setting, formatMsg)
-	}
-
-	var (
-		pluginID     = parsed[0]
-		settingKey   = parsed[1]
-		settingValue = parsed[2]
-	)
-
-	// Clean the strings. The setting value is allowed to be case
-	// sensitive.
-	pluginID = strings.ToLower(strings.TrimSpace(pluginID))
-	settingKey = strings.ToLower(strings.TrimSpace(settingKey))
-	settingValue = strings.TrimSpace(settingValue)
-
-	// Handle multiple values
-	if isMulti {
-		// Parse values
-		values := regexpPluginSettingMulti.FindString(setting)
-
-		// Verify the values are formatted as valid JSON
-		var s []string
-		err := json.Unmarshal([]byte(values), &s)
-		if err != nil {
-			return "", nil, err
-		}
-
-		// Re-encode the JSON. This will remove any funny
-		// formatting like whitespaces.
-		b, err := json.Marshal(s)
-		if err != nil {
-			return "", nil, err
-		}
-
-		// Save the value
-		settingValue = string(b)
-	}
-
-	return pluginID, &plugin.Setting{
-		Key:   settingKey,
-		Value: settingValue,
-	}, nil
-}
