+++ politeiawww/csrf.go
@@ -1,0 +1,5 @@
+// Copyright (c) 2022 The Decred developers
+// Use of this source code is governed by an ISC
+// license that can be found in the LICENSE file.
+
+package main
